release v0.1.0
this fixes an issue with latest AppSuite development version (that is to
be released as 7.8.0, later).

It also marks the first more stable release, leaving the 0.0 phase of
development. The plugin has been proven to work fine for quite some time
now. This release allows more easy bug-fixing release through bugfix
releases (as defined for semantic versioning).